// sdks/typescript/src/generated/CreateFinancialMandateRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type CreateFinancialMandateRequest = { id?: string, version?: number, principal_id: string, scope: Record<string, unknown> | null, metadata: Record<string, unknown> | null, starts_at?: string, expires_at?: string, };